Update to v2.1.4
Anki now checks for Python3.6, which we don't have in the
runtime. Drop that check and also tweak the one piece of
code that we know is not compatible with Python 3.5.

We'll look to upgrade to the new KDE runtime based
on freedesktop 18.08 as soon as possible, Python 3.6 will
be included there.

https://github.com/flathub/net.ankiweb.Anki/issues/3