Update Anki to 25.02
Since anki now uses yarn 4, some workarounds are required.

The flatpak environment does not include yarn 4, so it is added to the
sources.

flatpak-builder-tools upstream does not support yarn >=2, but there is
an open PR to add support [1]. That gets most of the way there, but
there is still a minor issue that I assume was introduced between yarn 2
and 4 not included in that PR, so I forked from that submitted a PR to
the developer of that PR [2].

Anki itself also needs minor changes (yarn-4-fixes.patch) since yarn 4
removed the --offline and --ignore-scripts options, but anki's offline
build still uses them. They are now instead set by the
flatpak-builder-tools script. Anki uses the YARN_BINARY environment
variable, but yarn will interpret any environment variables starting
with YARN_ as yarn configuration, so that will need to be renamed.

[1] https://github.com/flatpak/flatpak-builder-tools/pull/252
[2] https://github.com/catsout/flatpak-builder-tools/pull/1