WIP: Update Anki to 25.07.2
This updates the cargo and yarn dependencies, and begins work on python.
Anki switched to using uv, which means now we have to build that.

What still needs to be done is updating python3-modules.json. The first
issue I ran into is orjson being written in rust, meaning it has its own
cargo dependencies that will need to be included. That's where I left
off.